+++ src/request.ts
@@ -8,6 +8,8 @@ export type RequestType =
   | ":case-split"
   | ":docs-for"
   // | ":elaborate-term"
+  | ":generate-def"
+  | ":generate-def-next"
   // | ":hide-term-implicits"
   | ":interpret"
   | ":load-file"
@@ -18,6 +20,7 @@ export type RequestType =
   // | ":normalise-term"
   | ":print-definition"
   | ":proof-search"
+  | ":proof-search-next"
   | ":repl-completions"
   // | ":show-term-implicits"
   | ":type-of"
@@ -78,6 +81,18 @@ export namespace Request {
     type: ":docs-for"
   }
 
+  export interface GenerateDef extends RequestBase {
+    id: number
+    line: number
+    name: string
+    type: ":generate-def"
+  }
+
+  export interface GenerateDefNext extends RequestBase {
+    id: number
+    type: ":generate-def-next"
+  }
+
   export interface Interpret extends RequestBase {
     expression: string
     id: number
@@ -132,6 +147,11 @@ export namespace Request {
     type: ":proof-search"
   }
 
+  export interface ProofSearchNext extends RequestBase {
+    id: number
+    type: ":proof-search-next"
+  }
+
   export interface ReplCompletions extends RequestBase {
     id: number
     name: string
@@ -163,6 +183,8 @@ export type Request =
   | Request.CallsWho
   | Request.CaseSplit
   | Request.DocsFor
+  | Request.GenerateDef
+  | Request.GenerateDefNext
   | Request.Interpret
   | Request.LoadFile
   | Request.MakeCase
@@ -171,6 +193,7 @@ export type Request =
   | Request.Metavariables
   | Request.PrintDefinition
   | Request.ProofSearch
+  | Request.ProofSearchNext
   | Request.ReplCompletions
   | Request.TypeOf
   | Request.Version
@@ -186,6 +209,7 @@ const serialiseArgs = (req: Request): string => {
     case ":add-clause":
     case ":add-missing":
     case ":case-split":
+    case ":generate-def":
     case ":make-case":
     case ":make-lemma":
     case ":make-with":
@@ -210,10 +234,33 @@ const serialiseArgs = (req: Request): string => {
       return `${req.type} ${req.width}`
     case ":proof-search":
       return `${req.type} ${req.line} "${req.name}" (${req.hints.join(" ")})`
+    case ":generate-def-next":
+    case ":proof-search-next":
     case ":version":
       return req.type
   }
 }
 
+/**
+ * Serialise request according to the Idris 1 IDE protocol. Arguments, including
+ * the request type are always an s-expression list.
+ */
 export const serialiseRequest = (req: Request): string =>
   prependLen(`((${serialiseArgs(req)}) ${req.id})\n`)
+
+/**
+ * The Idris 2 IDE protocol introduces some new commands, and at least one
+ * breaking change to the serialised request format. Specifically commands
+ * without arguments are now simply atoms, rather than single-element lists.
+ */
+export const serialiseRequestV2 = (req: Request): string => {
+  if (
+    req.type === ":generate-def-next" ||
+    req.type === ":proof-search-next" ||
+    req.type === ":version"
+  ) {
+    return prependLen(`(${serialiseArgs(req)} ${req.id})\n`)
+  } else {
+    return prependLen(`((${serialiseArgs(req)}) ${req.id})\n`)
+  }
+}