+++ test/client/expected.ts
@@ -768,3 +768,26 @@ export const whoCalls: FinalReply.WhoCalls = {
   ok: true,
   type: ":return",
 }
+
+// Idris 2 only.
+export const generateDef: FinalReply.GenerateDef = {
+  def: "append [] ys = ys\nappend (x :: xs) ys = x :: append xs ys",
+  id: 20,
+  ok: true,
+  type: ":return",
+}
+
+export const generateDefNext: FinalReply.GenerateDef = {
+  def:
+    "append [] ys = ys\nappend (x :: xs) [] = x :: append xs []\nappend (x :: xs) (y :: ys) = x :: append xs (y :: ys)",
+  id: 21,
+  ok: true,
+  type: ":return",
+}
+
+export const proofSearchNext: FinalReply.ProofSearch = {
+  id: 22,
+  ok: true,
+  solution: "1",
+  type: ":return",
+}